[dotnet] [test] Start test webserver globally before any tests in Support project
