+++ dotnet/test/support/AssemblyFixture.cs
@@ -1,0 +1,55 @@
+// <copyright file="AssemblyFixture.cs" company="Selenium Committers">
+// Licensed to the Software Freedom Conservancy (SFC) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The SFC licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+// </copyright>
+
+using System.Threading.Tasks;
+using NUnit.Framework;
+using OpenQA.Selenium.Environment;
+
+namespace OpenQA.Selenium.Support;
+
+[SetUpFixture]
+public class AssemblyFixture
+{
+    public AssemblyFixture()
+    {
+    }
+
+    [OneTimeSetUp]
+    public async Task RunBeforeAnyTestAsync()
+    {
+        Internal.Logging.Log.SetLevel(Internal.Logging.LogEventLevel.Trace);
+
+        await EnvironmentManager.Instance.WebServer.StartAsync();
+        if (EnvironmentManager.Instance.Browser == Browser.Remote)
+        {
+            await EnvironmentManager.Instance.RemoteServer.StartAsync();
+        }
+    }
+
+    [OneTimeTearDown]
+    public async Task RunAfterAnyTestsAsync()
+    {
+        EnvironmentManager.Instance.CloseCurrentDriver();
+        await EnvironmentManager.Instance.WebServer.StopAsync();
+        if (EnvironmentManager.Instance.Browser == Browser.Remote)
+        {
+            await EnvironmentManager.Instance.RemoteServer.StopAsync();
+        }
+    }
+}
